requirements.txt
arrow==0.14.6
boto3==1.9.249
botocore==1.12.249
cached-property==1.5.1
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
colorlog==4.0.2
cycler==0.10.0
decorator==4.4.0
docutils==0.15.2
future==0.18.0
google-auth==1.6.3
humanfriendly==4.18
idna==2.8
jmespath==0.9.4
jsonpickle==1.2
kiwisolver==1.1.0
kubernetes==10.0.1
matplotlib==3.1.1
mypy-extensions==0.4.2
numpy==1.17.2
oauthlib==3.1.0
parsedatetime==2.4
protobuf==3.10.0
py==1.8.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyformance==0.4
pyparsing==2.4.2
PyStaticConfiguration==0.10.4
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
requests-oauthlib==1.2.0
retry==0.9.2
rsa==4.0
s3transfer==0.2.1
setuptools==41.4.0
signalfx==1.1.1
simplejson==3.16.0
six==1.12.0
sortedcontainers==2.1.0
sseclient-py==1.7
typing-extensions==3.7.4
urllib3==1.25.6
websocket-client==0.56.0
ws4py==0.5.1
